move portmapper files
