+++ net-tools/portmapper/src/current_mapping.rs
