+++ net-tools/portmapper/src/mapping.rs
